ActionのテストをServiceのテストへ移動
